Apps2pay/MINT/MINT_Loket/requirements.txt
Appium-Python-Client
selenium
pytest==7.1.2
allure-pytest==2.9.45
requests
pytest-xdist
pytest-html